backend/integrations/mcp-s7-server/requirements.txt
mcp
python-snap7
PyYAML
python-dotenv